+++ mlf_core/create/templates/mlflow/mlflow_tensorflow/{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug_no_hyphen }}/{{ cookiecutter.project_slug_no_hyphen }}.py
@@ -64,7 +64,7 @@ def start_training():
 
     with mlflow.start_run():
         # Enable the logging of all parameters, metrics and models to mlflow and Tensorboard
-        mlflow.tensorflow.autolog()
+        mlflow.autolog(1)
 
         # Log hardware and software
         MLFCore.log_sys_intel_conda_env()