+++ mlf_core/create/templates/mlflow/mlflow_xgboost_dask/{{ cookiecutter.project_slug }}/{{ cookiecutter.project_slug_no_hyphen }}/{{ cookiecutter.project_slug_no_hyphen }}.py
@@ -64,7 +64,7 @@ def start_training():
 
     with mlflow.start_run():
         # Enable the logging of all parameters, metrics and models to mlflow and Tensorboard
-        mlflow.xgboost.autolog()
+        mlflow.autolog(1)
 
         # Setup a Dask cluster to facilitate multiCPU/multiGPU training
         if use_cuda: