+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * [ALL TEMPLATES] Using new mlf-core/base:1.2.0 container, which is based on CUDA 11.2.1 and cudnn 8.1
 * [PYTORCH] Upgraded Pytorch to 1.7.1
 * [PYTORCH] Added set_deterministic
+* [ALL TEMPLATES] Using new mlflow autolog
+* [ALL TEMPLATES] Changed mlflow autolog for loss to every 1 iteration
 
 **Fixed**